+++ client/src/components/Review/ReviewList/ReviewListItem/ReviewListItem.tsx
@@ -3,10 +3,10 @@ import styled from 'styled-components';
 import { Mock } from '../../../../containers/ReviewContainer';
 import formatDate from '../../../../utils/formatDate';
 import ReviewDetail from './ReviewDetail/ReviewDetail';
+import ReviewSummary from './ReviewSummary/ReviewSummary';
 import { generateStars } from './Star';
-import CHEVRON_DOWN from './chevronDown.png';
 
-const MAX_TITLE_WIDTH = 700;
+export const MAX_TITLE_WIDTH = 700;
 
 const Container = styled.li`
   padding: 10px 20px;
@@ -27,41 +27,6 @@ const ReviewStars = styled.div`
   width: 80px;
   margin-right: 40px;
   display: flex;
-`;
-
-type ClickContainerProps = {
-  isClickable: boolean;
-};
-const ClickContainer = styled.div<ClickContainerProps>`
-  flex-grow: 1;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  cursor: ${(props) => (props.isClickable ? 'pointer' : 'default')};
-`;
-
-const ReviewTitle = styled.span`
-  font-size: ${(props) => props.theme.fontSize.small};
-  max-width: ${MAX_TITLE_WIDTH}px;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-`;
-
-type SeeMoreProps = {
-  isClose: boolean;
-};
-const SeeMore = styled.button<SeeMoreProps>`
-  width: 10px;
-  height: 10px;
-  margin-right: 16px;
-  cursor: pointer;
-  border: none;
-  background-color: transparent;
-  background-image: url(${CHEVRON_DOWN});
-  background-size: contain;
-  transform: rotate(${(props) => (props.isClose ? '-180' : '0')}deg);
-  transition: transform 0.3s ease-in-out;
 `;
 
 const ReviewDate = styled.div`
@@ -96,10 +61,13 @@ const ReviewListItem = (props: Props): JSX.Element => {
     <Container>
       <ReviewDisplayContainer>
         <ReviewStars>{stars}</ReviewStars>
-        <ClickContainer onClick={handleReviewItemClick} isClickable={hasMoreContent}>
-          <ReviewTitle ref={titleRef}>{review.content}</ReviewTitle>
-          {hasMoreContent && <SeeMore isClose={reviewDetailOpen} />}
-        </ClickContainer>
+        <ReviewSummary
+          content={review.content}
+          onClick={handleReviewItemClick}
+          isClickable={hasMoreContent}
+          reviewTitleRef={titleRef}
+          reviewDetailOpen={reviewDetailOpen}
+        />
         <ReviewDate>{formatDate(review.updatedAt)}</ReviewDate>
       </ReviewDisplayContainer>
       {reviewDetailOpen && hasMoreContent && <ReviewDetail review={review} />}
